BugFix: Fixed error on editing the product
Added missing fields to product edit screen.
Also removed unecessary Update method on product class, the way it was
done we had to add and remove fields from that method manually
every time something changed on the model. Since we are using entity
framework we can just update using the object directly, so new or
deleted fields will be handled by the framework.